requirements.txt
Django
django-filter
django-nose==1.4.6
djangorestframework
djangorestframework-gis
protobuf==3.15.4
psycopg2-binary>=2.8
gunicorn==19.9.0
drf-yasg
django-rest-swagger
datapackage
requests